refactor updateMergeRequest action
