+++ lib/helpers/index.js
@@ -3,9 +3,15 @@
 var getSettings = require('./getSettings'),
 	isSuitableBuild = require('./isSuitableBuild'),
 	renderTemplate = require('./renderTemplate'),
+	stringifySection = require('./stringifySection'),
+	hasSection = require('./hasSection'),
+	replaceSection = require('./replaceSection'),
 	makeBuildInfoComment = require('./makeBuildInfoComment');
 
 exports.getSettings = getSettings;
 exports.isSuitableBuild = isSuitableBuild;
 exports.renderTemplate = renderTemplate;
+exports.stringifySection = stringifySection;
+exports.hasSection = hasSection;
+exports.replaceSection = replaceSection;
 exports.makeBuildInfoComment = makeBuildInfoComment;
